requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.15.0
mypy_extensions==1.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
ruff==0.11.11
SQLAlchemy==2.0.40
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
